Remove default mappings
These mappings are not necessarily going to be ingrained for
non-thoughtbotters.

I know for a fact that even within thoughtbot not everyone uses them:
take @rook's [dotfiles] for example. I also use `<Leader>o` for the
equivalent of `RunNearestSpec` in my own workflow.

[dotfiles]: https://github.com/r00k/dotfiles/blob/master/vimrc#L79

I've moved the previous Default mappings to the README as Recommended
mappings.