Remove explicit Dispatch support
Test runners and pre-loaders that require a different command to be run
should be enabled by setting `g:rspec_command` in your .vimrc. The
pre-existing support for Dispatch didn't work if you loaded vim-rspec
before Dispatch. Rather than work around this or any other
runner/pre-loader issues, we should leave it up to the user to set the
command.